Reorganize duckchat module
