+++ duckchat/duckchat-impl/src/main/AndroidManifest.xml
@@ -18,12 +18,12 @@
 
     <application>
         <activity
-            android:name="com.duckduckgo.duckchat.impl.DuckChatSettingsActivity"
+            android:name="com.duckduckgo.duckchat.impl.ui.DuckChatSettingsActivity"
             android:exported="false"
             android:label="@string/duck_chat_title"
             android:parentActivityName="com.duckduckgo.app.settings.SettingsActivity" />
         <activity
-            android:name="com.duckduckgo.duckchat.impl.DuckChatWebViewActivity"
+            android:name="com.duckduckgo.duckchat.impl.ui.DuckChatWebViewActivity"
             android:exported="false"
             android:label="@string/duck_chat_title"
             android:configChanges="keyboardHidden|orientation|screenSize|smallestScreenSize|screenLayout|navigation|keyboard"/>
