+++ duckchat/duckchat-impl/src/main/java/com/duckduckgo/duckchat/impl/di/DuckChatDataStoreModule.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.duckduckgo.duckchat.impl
+package com.duckduckgo.duckchat.impl.di
 
 import android.content.Context
 import androidx.datastore.core.DataStore
