+++ duckchat/duckchat-impl/src/main/res/layout/activity_duck_chat_settings.xml
@@ -19,7 +19,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".DuckChatSettingsActivity">
+    tools:context=".ui.DuckChatSettingsActivity">
 
     <include
         android:id="@+id/includeToolbar"
